requirements.txt
streamlit
pandas
plotly
google-api-python-client
google-auth